Initial segment code

+++ cli2c/segment/ht16k33-segment.h
@@ -48,7 +48,7 @@ void            HT16K33_clear_buffer(void);
 void            HT16K33_set_brightness(uint8_t brightness);
 void            HT16K33_set_number(uint8_t number, uint8_t digit, bool has_dot);
 void            HT16K33_set_glyph(uint8_t glyph, uint8_t digit, bool has_dot);
-void            HT16K33_show_value(int16_t value, bool decimal);
+void            HT16K33_show_value(int value, bool decimal);
 
 static uint32_t bcd(uint32_t base);
 static void     HT16K33_sleep_ms(int ms);
